requirements.txt
scipy
pytictoc
torchmetrics